+++ tests/Concerns/DefineRoutes.php
@@ -33,6 +33,18 @@ trait DefineRoutes
             ], $status);
         })->name('status');
 
+        $router->get('header', function () {
+            return response('', 200, [
+                'foo' => 'bar',
+            ]);
+        })->name('header');
+
+        $router->get('session', function () {
+            session(['foo' => 'bar']);
+
+            return response();
+        })->name('session');
+
         $router->get('download/{filename}', function ($filename) {
             Storage::put($filename, 'test');
 
@@ -66,15 +78,39 @@ trait DefineRoutes
 
         $router->get('json', function () {
             return response()->json([
-               'foo' => [
-                   'bar' => 'baz',
-               ],
-                'qux' => 1,
-               'quuz' => [
-                   'corge',
-                   'grault',
-               ],
-           ]);
+                'foo'  => [
+                    'bar' => 'baz',
+                ],
+                'qux'  => 1,
+                'quuz' => [
+                    'corge',
+                    'grault',
+                ],
+                'foobar' => [
+                    'foobar_foo' => 'foo',
+                    'foobar_bar' => 'bar',
+                ],
+                '0'    => ['foo'],
+                'bars' => [
+                    ['bar' => 'foo 0', 'foo' => 'bar 0'],
+                    ['bar' => 'foo 1', 'foo' => 'bar 1'],
+                    ['bar' => 'foo 2', 'foo' => 'bar 2'],
+                ],
+                'baz' => [
+                    ['foo' => 'bar 0', 'bar' => ['foo' => 'bar 0', 'bar' => 'foo 0']],
+                    ['foo' => 'bar 1', 'bar' => ['foo' => 'bar 1', 'bar' => 'foo 1']],
+                ],
+                'barfoo' => [
+                    ['bar' => ['bar' => 'foo 0']],
+                    ['bar' => ['bar' => 'foo 0', 'foo' => 'foo 0']],
+                    ['bar' => ['foo' => 'bar 0', 'bar' => 'foo 0', 'rab' => 'rab 0']],
+                ],
+                'numeric_keys' => [
+                    2 => ['bar' => 'foo 0', 'foo' => 'bar 0'],
+                    3 => ['bar' => 'foo 1', 'foo' => 'bar 1'],
+                    4 => ['bar' => 'foo 2', 'foo' => 'bar 2'],
+                ],
+            ]);
         });
 
         $router->post('validate', function (Request $request) {