+++ tests/Expect/Response/toHaveExactJson.php
@@ -5,13 +5,37 @@ use PHPUnit\Framework\ExpectationFailedException;
 
 test('pass', function () {
     expect(get('json'))->toHaveExactJson([
-        'foo' => [
+        'foo'  => [
             'bar' => 'baz',
         ],
         'qux'  => 1,
         'quuz' => [
             'corge',
             'grault',
+        ],
+        'foobar' => [
+            'foobar_foo' => 'foo',
+            'foobar_bar' => 'bar',
+        ],
+        '0'    => ['foo'],
+        'bars' => [
+            ['bar' => 'foo 0', 'foo' => 'bar 0'],
+            ['bar' => 'foo 1', 'foo' => 'bar 1'],
+            ['bar' => 'foo 2', 'foo' => 'bar 2'],
+        ],
+        'baz' => [
+            ['foo' => 'bar 0', 'bar' => ['foo' => 'bar 0', 'bar' => 'foo 0']],
+            ['foo' => 'bar 1', 'bar' => ['foo' => 'bar 1', 'bar' => 'foo 1']],
+        ],
+        'barfoo' => [
+            ['bar' => ['bar' => 'foo 0']],
+            ['bar' => ['bar' => 'foo 0', 'foo' => 'foo 0']],
+            ['bar' => ['foo' => 'bar 0', 'bar' => 'foo 0', 'rab' => 'rab 0']],
+        ],
+        'numeric_keys' => [
+            2 => ['bar' => 'foo 0', 'foo' => 'bar 0'],
+            3 => ['bar' => 'foo 1', 'foo' => 'bar 1'],
+            4 => ['bar' => 'foo 2', 'foo' => 'bar 2'],
         ],
     ]);
 });
@@ -35,13 +59,37 @@ test('pass negated', function () {
 
 test('fails negated', function () {
     expect(get('json'))->not->toHaveExactJson([
-        'foo' => [
+        'foo'  => [
             'bar' => 'baz',
         ],
         'qux'  => 1,
         'quuz' => [
             'corge',
             'grault',
+        ],
+        'foobar' => [
+            'foobar_foo' => 'foo',
+            'foobar_bar' => 'bar',
+        ],
+        '0'    => ['foo'],
+        'bars' => [
+            ['bar' => 'foo 0', 'foo' => 'bar 0'],
+            ['bar' => 'foo 1', 'foo' => 'bar 1'],
+            ['bar' => 'foo 2', 'foo' => 'bar 2'],
+        ],
+        'baz' => [
+            ['foo' => 'bar 0', 'bar' => ['foo' => 'bar 0', 'bar' => 'foo 0']],
+            ['foo' => 'bar 1', 'bar' => ['foo' => 'bar 1', 'bar' => 'foo 1']],
+        ],
+        'barfoo' => [
+            ['bar' => ['bar' => 'foo 0']],
+            ['bar' => ['bar' => 'foo 0', 'foo' => 'foo 0']],
+            ['bar' => ['foo' => 'bar 0', 'bar' => 'foo 0', 'rab' => 'rab 0']],
+        ],
+        'numeric_keys' => [
+            2 => ['bar' => 'foo 0', 'foo' => 'bar 0'],
+            3 => ['bar' => 'foo 1', 'foo' => 'bar 1'],
+            4 => ['bar' => 'foo 2', 'foo' => 'bar 2'],
         ],
     ]);
 })->throws(ExpectationFailedException::class);